Separate load operation into a new `CacheFeedImageDataUseCaseTests`
