+++ EssentialFeed/EssentialFeedTests/Feed Cache/LocalFeedImageLoaderTests/CacheFeedImageDataUseCaseTests.swift
@@ -1,0 +1,40 @@
+//
+//  CacheFeedImageDataUseCaseTests.swift
+//  EssentialFeed
+//
+//  Created by Anthony on 13/4/25.
+//
+
+import Foundation
+import XCTest
+import EssentialFeed
+
+final class CacheFeedImageDataUseCaseTests: XCTestCase {
+    func test_init_doesNotMessageStoreUponCreation() {
+        let (_, store) = makeSUT()
+        
+        XCTAssertEqual(store.receivedMessages, [])
+    }
+    
+    func test_saveImageDataForURL_requestsImageDataInsertionForURL() {
+        let (sut, store) = makeSUT()
+        let url = anyURL()
+        let data = anydata()
+        
+        sut.save(data, for: url) { _ in }
+        
+        XCTAssertEqual(store.receivedMessages, [.insert(dataFor: url)])
+    }
+    
+    // MARK: - Helpers
+    private func makeSUT(file: StaticString = #file, line: UInt = #line) -> (sut: LocalFeedImageDataLoader, store: FeedImageDataStoreSpy) {
+        
+        let store = FeedImageDataStoreSpy()
+        let sut = LocalFeedImageDataLoader(store: store)
+        
+        trackMemoryLeaks(store, file: file, line: line)
+        trackMemoryLeaks(sut, file: file, line: line)
+        
+        return (sut, store)
+    }
+}
